- move functions onClick and isFormSubmitted to another file, because utilizing in both search results components
