+++ src/components/Personality/PersonalitySearchResultSection.tsx
@@ -1,9 +1,6 @@
 import { Row } from "antd";
-import { useTranslation } from "next-i18next";
-import { useRouter } from "next/router";
 import React from "react";
 
-import api from "../../api/personality";
 import Label from "../Label";
 import PersonalityCard from "./PersonalityCard";
 
@@ -11,28 +8,10 @@ const PersonalitySearchResultSection = ({
     personalities,
     label,
     selectPersonality,
+    isFormSubmitted,
+    onClick,
 }) => {
-    const { t } = useTranslation();
-    const router = useRouter();
-
     const isCreatingClaim = selectPersonality !== null;
-
-    const createPersonality = async (personality) => {
-        const personalityCreated = await api.createPersonality(personality, t);
-        const { slug } = personalityCreated;
-        const newPersonality = {
-            ...personality,
-            ...personalityCreated,
-        };
-        const createClaim = () => {
-            selectPersonality(newPersonality);
-        };
-
-        // Redirect to personality list in case _id is not present
-        const path = slug ? `/personality/${slug}` : "/personality";
-
-        isCreatingClaim ? createClaim() : router.push(path);
-    };
 
     return personalities.length ? (
         <Row
@@ -53,8 +32,9 @@ const PersonalitySearchResultSection = ({
                             summarized={true}
                             enableStats={false}
                             hrefBase="./"
-                            onClick={createPersonality}
+                            onClick={onClick}
                             key={i}
+                            isFormSubmitted={isFormSubmitted}
                         />
                     )
             )}
